moved DTO classes to a separate package
