+++ ahaar/src/main/java/com/innovatrix/ahaar/model/ApplicationUser.java
@@ -1,5 +1,6 @@
 package com.innovatrix.ahaar.model;
 
+import com.innovatrix.ahaar.DTO.ApplicationUserDTO;
 import jakarta.persistence.*;
 import lombok.*;
